already bump to version 1.2.0; restructure files
