+++ definition.yml
@@ -1,5 +1,5 @@
 name: files
-version: 1.1.0
+version: 1.2.0
 author: [gynt, TheRedDaemon]
 dependencies:
   frontend: ">=1.0.2"
