+++ overrides/registry.lua
@@ -1,0 +1,8 @@
+--- Keys in this dictionary are lower case. Values have their original case.
+local FILE_OVERRIDES = {}
+local FILE_OVERRIDE_FUNCTIONS = {}
+
+return {
+  ["FILE_OVERRIDES"] = FILE_OVERRIDES,
+  ["FILE_OVERRIDE_FUNCTIONS"] = FILE_OVERRIDE_FUNCTIONS,
+}
